refactor: move providers to separate folder, move date-icon to ui folder
