+++ src/components/date-select.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import { formatDate, getNextWeek, getToday, getTomorrow } from '@/lib/utils';
 
-import { DateIcon } from './date-icon';
+import { DateIcon } from './ui/date-icon';
 import { Button, ButtonProps } from '@/components/ui/button';
 import { Calendar } from '@/components/ui/calendar';
 import {
